refactor: 공통으로 사용되는 BindingAdapter util 패키지로 이동
