+++ app/src/main/java/woowacourse/shopping/ui/products/ProductContentsActivity.kt
@@ -102,13 +102,3 @@ class ProductContentsActivity : AppCompatActivity() {
 }
 
 fun List<ProductUiModel>.isLoading() = this + LoadingUiModel + LoadingUiModel
-
-@BindingAdapter("imageUrl")
-fun ImageView.bindUrlToImage(imageUrl: String?) {
-    urlToImage(context, imageUrl)
-}
-
-@BindingAdapter("isVisible")
-fun View.setIsVisible(isVisible: Boolean) {
-    visibility = if (isVisible) View.VISIBLE else View.GONE
-}
