+++ app/src/main/java/woowacourse/shopping/ui/utils/BindingAdapters.kt
@@ -1,0 +1,15 @@
+package woowacourse.shopping.ui.utils
+
+import android.view.View
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter("imageUrl")
+fun ImageView.bindUrlToImage(imageUrl: String?) {
+    urlToImage(context, imageUrl)
+}
+
+@BindingAdapter("isVisible")
+fun View.setIsVisible(isVisible: Boolean) {
+    visibility = if (isVisible) View.VISIBLE else View.GONE
+}
